requirements.txt
selenium
beautifulsoup4
python-telegram-bot==13.15
beautifulsoup4
psycopg2-binary
